requirements.txt
Django>=4.2
djangorestframework==3.16.1
pytesseract
pillow
python-decouple
